Split debugging commands to new sub-cog.
